+++ dbo/Stored Procedures/DsmrApi_SelectLastData.sql
@@ -1,0 +1,20 @@
+﻿CREATE PROCEDURE [dbo].[DsmrApi_SelectLastData]
+	@sensorId INT
+AS
+BEGIN
+	SELECT TOP (1)
+       [PowerUsage]
+      ,[PowerProduction]
+      ,[EnergyUsage]
+      ,[EnergyProduction]
+      ,[Tariff]
+      ,[GasUsage]
+      ,[GasFlow]
+      ,[OutsideAirTemperature]
+      ,[Temperature]
+      ,[Pressure]
+      ,[RH]
+      ,[Timestamp]
+	FROM [dbo].[DataPoints]
+	ORDER BY [Timestamp] DESC
+END