+++ dbo/Stored Procedures/DsmrApi_SelectWeeklyHigh.sql
@@ -1,0 +1,13 @@
+﻿CREATE PROCEDURE [dbo].[DsmrApi_SelectWeeklyHigh]
+	@sensorId INT
+AS
+BEGIN 
+	SELECT MAX([PowerUsage]) AS [PowerUsage],
+		   MAX([PowerProduction]) AS [PowerProduction],
+		   MAX([GasFlow]) * 1000.0 AS [GasFlow], -- In liters/minute
+		   MAX([Temperature]) AS [Temperature],
+		   MAX([OutsideAirTemperature]) AS [Oat]
+	  FROM [dbo].[DataPoints]
+	  WHERE [SensorId] = @sensorId
+		AND [Timestamp] >= DATEADD(DAY, -7, GETUTCDATE())
+END